+++ kotlin/samples/tictactoe/android/src/main/java/com/squareup/sample/authworkflow/AuthorizingCoordinator.kt
@@ -19,8 +19,8 @@ import android.view.View
 import android.widget.TextView
 import com.squareup.coordinators.Coordinator
 import com.squareup.sample.tictactoe.R
-import com.squareup.viewregistry.LayoutBinding
-import com.squareup.viewregistry.ViewBinding
+import com.squareup.workflow.ui.LayoutBinding
+import com.squareup.workflow.ui.ViewBinding
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable