+++ kotlin/samples/tictactoe/android/src/main/java/com/squareup/sample/panel/PanelContainer.kt
@@ -25,8 +25,8 @@ import android.widget.FrameLayout
 import com.squareup.sample.mainworkflow.display
 import com.squareup.sample.mainworkflow.isTablet
 import com.squareup.sample.tictactoe.R
-import com.squareup.viewregistry.ModalContainer
-import com.squareup.viewregistry.ViewBinding
+import com.squareup.workflow.ui.ModalContainer
+import com.squareup.workflow.ui.ViewBinding
 import kotlin.math.min
 
 /**
@@ -56,6 +56,7 @@ internal class PanelBodyWrapper
   attributeSet: AttributeSet? = null
 ) : FrameLayout(context, attributeSet) {
   init {
+    @Suppress("DEPRECATION")
     background = ColorDrawable(resources.getColor(R.color.panelBody))
   }