+++ kotlin/samples/tictactoe/common/src/main/java/com/squareup/sample/gameworkflow/CompletedGame.kt
@@ -17,23 +17,20 @@ package com.squareup.sample.gameworkflow
 
 import com.squareup.workflow.Snapshot
 import com.squareup.workflow.parse
-import com.squareup.workflow.readByteStringWithLength
-import com.squareup.workflow.writeByteStringWithLength
 import okio.ByteString
 
 /**
  * The [lastTurn] of a tic tac toe game, and its [ending]. Serves as the
- * result type for [TakeTurnsReactor].
+ * result type for [TakeTurnsWorkflow].
  */
 data class CompletedGame(
   val ending: Ending,
-  val lastTurn: Turn
+  val lastTurn: Turn = Turn()
 ) {
 
   fun toSnapshot(): Snapshot {
     return Snapshot.write { sink ->
-      sink.writeInt(this.ending.ordinal)
-      sink.writeByteStringWithLength(this.lastTurn.toSnapshot().bytes)
+      sink.writeInt(ending.ordinal)
     }
   }
 
@@ -41,10 +38,7 @@ data class CompletedGame(
     fun fromSnapshot(byteString: ByteString): CompletedGame {
       return byteString.parse { source ->
         CompletedGame(
-            Ending.values()[source.readInt()],
-            Turn.fromSnapshot(
-                source.readByteStringWithLength()
-            )
+            Ending.values()[source.readInt()]
         )
       }
     }