+++ kotlin/workflow-rx2/src/test/java/com/squareup/workflow/rx2/WorkflowContextsTest.kt
@@ -17,20 +17,20 @@ package com.squareup.workflow.rx2
 
 import com.squareup.workflow.EventHandler
 import com.squareup.workflow.Snapshot
-import com.squareup.workflow.StatelessWorkflow
+import com.squareup.workflow.StatefulWorkflow
 import com.squareup.workflow.Workflow
 import com.squareup.workflow.WorkflowAction.Companion.emitOutput
 import com.squareup.workflow.WorkflowAction.Companion.enterState
 import com.squareup.workflow.WorkflowAction.Companion.noop
 import com.squareup.workflow.WorkflowContext
 import com.squareup.workflow.invoke
+import com.squareup.workflow.stateless
 import com.squareup.workflow.testing.testFromStart
 import io.reactivex.subjects.SingleSubject
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
-import kotlin.test.fail
 
 class WorkflowContextsTest {
 
@@ -41,7 +41,7 @@ class WorkflowContextsTest {
     val single = singleSubject
         .doOnSubscribe { subscriptions++ }
         .doOnDispose { disposals++ }
-    val workflow = StatelessWorkflow<String, Unit> { context ->
+    val workflow = Workflow.stateless<String, Unit> { context ->
       context.onSuccess(single) { emitOutput(it) }
     }
 
@@ -70,8 +70,11 @@ class WorkflowContextsTest {
     val single = singleSubject
         .doOnSubscribe { subscriptions++ }
         .doOnDispose { disposals++ }
-    val workflow = object : Workflow<Unit, Boolean, Nothing, Unit> {
-      override fun initialState(input: Unit): Boolean = true
+    val workflow = object : StatefulWorkflow<Unit, Boolean, Nothing, Unit>() {
+      override fun initialState(
+        input: Unit,
+        snapshot: Snapshot?
+      ): Boolean = true
 
       override fun compose(
         input: Unit,
@@ -85,7 +88,6 @@ class WorkflowContextsTest {
       }
 
       override fun snapshotState(state: Boolean): Snapshot = Snapshot.EMPTY
-      override fun restoreState(snapshot: Snapshot): Boolean = fail("not expected")
     }
 
     assertEquals(0, subscriptions)